changed MonitoringTag and new generated
