+++ docs/JobApi.md
@@ -5,7 +5,6 @@ All URIs are relative to *http://localhost:9000*
 Method | HTTP request | Description
 ------------- | ------------- | -------------
 [**addJob**](JobApi.md#addJob) | **POST** /jobs | 
-[**deleteSchedule**](JobApi.md#deleteSchedule) | **DELETE** /schedule/{id} | 
 [**findJob**](JobApi.md#findJob) | **GET** /jobs/{id} | 
 [**findJobs**](JobApi.md#findJobs) | **GET** /jobs | 
 [**jobGraph**](JobApi.md#jobGraph) | **GET** /jobs/{id}/graph | 
@@ -64,61 +63,6 @@ Name | Type | Description  | Notes
 ### HTTP request headers
 
  - **Content-Type**: application/json
- - **Accept**: application/json
-
-<a name="deleteSchedule"></a>
-# **deleteSchedule**
-> Queue deleteSchedule(id)
-
-
-
-Deletes the schedule identified by the given id and all corresponding processes and nodes. 
-
-### Example
-```java
-// Import classes:
-//import io.github.cloudiator.rest.ApiClient;
-//import io.github.cloudiator.rest.ApiException;
-//import io.github.cloudiator.rest.Configuration;
-//import io.github.cloudiator.rest.auth.*;
-//import io.github.cloudiator.rest.api.JobApi;
-
-ApiClient defaultClient = Configuration.getDefaultApiClient();
-
-// Configure API key authorization: ApiKeyAuth
-ApiKeyAuth ApiKeyAuth = (ApiKeyAuth) defaultClient.getAuthentication("ApiKeyAuth");
-ApiKeyAuth.setApiKey("YOUR API KEY");
-// Uncomment the following line to set a prefix for the API key, e.g. "Token" (defaults to null)
-//ApiKeyAuth.setApiKeyPrefix("Token");
-
-JobApi apiInstance = new JobApi();
-String id = "id_example"; // String | Unique identifier of the resource
-try {
-    Queue result = apiInstance.deleteSchedule(id);
-    System.out.println(result);
-} catch (ApiException e) {
-    System.err.println("Exception when calling JobApi#deleteSchedule");
-    e.printStackTrace();
-}
-```
-
-### Parameters
-
-Name | Type | Description  | Notes
-------------- | ------------- | ------------- | -------------
- **id** | **String**| Unique identifier of the resource |
-
-### Return type
-
-[**Queue**](Queue.md)
-
-### Authorization
-
-[ApiKeyAuth](../README.md#ApiKeyAuth)
-
-### HTTP request headers
-
- - **Content-Type**: Not defined
  - **Accept**: application/json
 
 <a name="findJob"></a>
