+++ docs/Schedule.md
@@ -7,6 +7,7 @@ Name | Type | Description | Notes
 **job** | **String** | The identifier of the job |  [optional]
 **instantiation** | [**InstantiationEnum**](#InstantiationEnum) | If the instantiation should be handled AUTOMATIC or MANUAL |  [optional]
 **id** | **String** |  |  [optional]
+**owner** | **String** |  |  [optional]
 **processes** | [**List&lt;CloudiatorProcess&gt;**](CloudiatorProcess.md) |  |  [optional]
