+++ src/main/java/io/github/cloudiator/rest/model/Schedule.java
@@ -91,6 +91,9 @@ public class Schedule implements Serializable {
   @SerializedName("id")
   private String id = null;
 
+  @SerializedName("owner")
+  private String owner = null;
+
   @SerializedName("processes")
   private List<CloudiatorProcess> processes = null;
 
@@ -148,6 +151,24 @@ public class Schedule implements Serializable {
     this.id = id;
   }
 
+  public Schedule owner(String owner) {
+    this.owner = owner;
+    return this;
+  }
+
+   /**
+   * Get owner
+   * @return owner
+  **/
+  @ApiModelProperty(value = "")
+  public String getOwner() {
+    return owner;
+  }
+
+  public void setOwner(String owner) {
+    this.owner = owner;
+  }
+
   public Schedule processes(List<CloudiatorProcess> processes) {
     this.processes = processes;
     return this;
@@ -187,12 +208,13 @@ public class Schedule implements Serializable {
     return Objects.equals(this.job, schedule.job) &&
         Objects.equals(this.instantiation, schedule.instantiation) &&
         Objects.equals(this.id, schedule.id) &&
+        Objects.equals(this.owner, schedule.owner) &&
         Objects.equals(this.processes, schedule.processes);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(job, instantiation, id, processes);
+    return Objects.hash(job, instantiation, id, owner, processes);
   }
 
 
@@ -204,6 +226,7 @@ public class Schedule implements Serializable {
     sb.append("    job: ").append(toIndentedString(job)).append("\n");
     sb.append("    instantiation: ").append(toIndentedString(instantiation)).append("\n");
     sb.append("    id: ").append(toIndentedString(id)).append("\n");
+    sb.append("    owner: ").append(toIndentedString(owner)).append("\n");
     sb.append("    processes: ").append(toIndentedString(processes)).append("\n");
     sb.append("}");
     return sb.toString();
